Changes to better support Swift REPL
