+++ Sources/Run/main.swift
@@ -1,23 +1,9 @@
 import App
-import Service
-import Vapor
 import Foundation
 
 // The contents of main are wrapped in a do/catch block because any errors that get raised to the top level will crash Xcode
 do {
-    var config = Config.default()
-    var env = try Environment.detect()
-    var services = Services.default()
-
-    try App.configure(&config, &env, &services)
-
-    let app = try Application(
-        config: config,
-        environment: env,
-        services: services
-    )
-
-    try App.boot(app)
+    let app = try App.bootstrap()
 
     try app.run()
 } catch {
